build: use pkg-config to find OpenSSL.
The pkg-config file already relied on OpenSSL's pkg-config file to be
present, and there is no reason to expect a modern system not to have
OpenSSL (plus, if the pkg-config files are not present, pkg-config allows
overriding the path without having to resort to --with-openssl).

This also removes forced inclusion of OPENSSL flags in login_duo and
pam_duo that do not make use of OpenSSL directly.