Const correctness for various implementations of PAM
Variables that will be passed to a pam_* function should be marked as
duopam_const to account for different implementation. For example, Linux
distros include the const qualifier, whereas AIX and Solaris based OSs do
not include the qualifier.

Variables that won't be passed to a pam_* function should unconditionally be
const.