properly return PAM_MAXTRIES if prompts=N set where N != 3
The auth loop limits at cfg.prompts, but the check at the end of the loop for whether to return PAM_MAXTRIES compares to the constand MAX_PROMPTS instead. This works if prompts=3 in pam_duo.conf, or isn't specified (same effect), but fails if any other value is used, returning PAM_SERVICE_ERR instead.