While Duo can be configured to use an HTTP proxy, the authlogin_duo SELinux module only permits outbound connections to TCP ports with the http_port_t SELinux label.
This is a problem, because HTTP proxies are typically configured to
listen on ports with the http_cache_port_t SELinux label.  (For
example, 8080 is a common port to use for HTTP proxies, and port 8080
has the http_cache_port_t SELinux label, not the http_port_t label.)

This change updates the authlogin_duo SELinux module to permit the
sshd_t type to connect to both http_port_t and http_cache_port_t
ports, and increments the module version number.