Make SELinux support selectable.
Specifically, instead of automatically building the SELinux module if
the necessary components are available, use the --with-selinux
configure option to enable SELinux support.  By default, SELinux
support will not be enabled.  If --with-selinux is specified, but the
necessary development tools to build the SELinux module are not
available, configure will fail.