Feature: add remember me to pam_duo
This will add a remember me feature to pam_duo. By simply adding
a "remember_me" configuration setting to pam_duo.conf such as:

remember_me = 14D

when a user authenticates with duo from a host the connection
authentication for that host will be remembered for that specified
number of seconds, minutes, hours or days using a value such as
10s or 20m or 12h or 30d to specify durations of 10 seconds or
20 minutes or 12 hours or 30 days.

subsequent connections from that host will not require duo authentication
until the time period expires.

If the remember me feature is enabled and duo confirmation is performed
the user will be prompted to select whether the authentication for that
host should be remembered or not.

The remember me information is stored in a binary file called
~/.pam_duo_remember on the server side of the connection. An optional
remember_me_protected option in pam_duo.conf can be set to yes if it is
desired to have these ~/.pam_duo_remember files to be owned by root instead
of the user account.