Use IP Address from TEST-NET-1 range
The duo_local_ip function creates a UDP socket to a
public server to determine the right outgoing interface
and source IP address. The function currently depends
on Google's 8.8.8.8 DNS server for requests.

However, the same experience can be achieved by using
any of the IP addresses reserved for documentation in
RFC 5737 Section 3 [1]

  > The blocks 192.0.2.0/24 (TEST-NET-1),
  > 198.51.100.0/24 (TEST-NET-2), and
  > 203.0.113.0/24 (TEST-NET-3) are provided for use in
  >  documentation

This PR updates the cod to use the address 192.0.2.1
which is reserved (instead of the current 8.8.8.8)  and
works in determining the outgoing interface and source
IP address.

[1]: https://datatracker.ietf.org/doc/html/rfc5737#section-3

Signed-off-by: Raymond Etornam <retornam@users.noreply.github.com>